// src/data/node-types/car-models/types/ApiCarModelHasImageRelationship.ts
import type {ApiNodeType} from "../../../types/ApiNodeType"
import type {ApiImageNode} from "../../images/types/ApiImageNode"

export type ApiCarModelHasImageRelationship = {
    data: {
        relationship_id: number
        relationship_name: 'has-image'
        partner_node: {
            node_type: ApiNodeType.IMAGE
            data: ApiImageNode['attributes']
        }
        created_at: string
        updated_at: string
    }
}